fix: moving calc_age to utils
