+++ sn_interface/src/network_knowledge/node_info.rs
@@ -6,7 +6,7 @@
 // KIND, either express or implied. Please review the Licences for the specific language governing
 // permissions and limitations relating to use of the SAFE Network Software.
 
-use crate::types::{calc_age, Peer, PublicKey};
+use crate::types::{utils::calc_age, Peer, PublicKey};
 use ed25519_dalek::Keypair;
 use std::{
     fmt::{self, Display, Formatter},
